+++ example/src/App.tsx
@@ -5,10 +5,11 @@
 import React from 'react';
 import { Platform, SafeAreaView, ScrollView, StyleSheet, Switch, Text, View, TouchableOpacity } from 'react-native';
 import type { MSALResult, MSALWebviewParams } from 'react-native-msal';
-import B2CClient from './b2cClient';
+
+import { B2CClient } from './b2cClient';
 import { b2cConfig, b2cScopes as scopes } from './msalConfig';
 
-const b2cClient = new B2CClient(b2cConfig, false);
+const b2cClient = new B2CClient(b2cConfig);
 
 export default function App() {
   const [authResult, setAuthResult] = React.useState<MSALResult | null>(null);
@@ -43,7 +44,7 @@ export default function App() {
 
   const handleAcquireTokenPress = async () => {
     try {
-      const res = await b2cClient.acquireTokenSilent({ scopes });
+      const res = await b2cClient.acquireTokenSilent({ scopes, forceRefresh: true });
       setAuthResult(res);
     } catch (error) {
       console.warn(error);