+++ src/types.ts
@@ -1,12 +1,53 @@
 import type { Configuration } from '@azure/msal-browser';
 
 export interface IPublicClientApplication {
-  init(): Promise<void>;
-  acquireToken(params: MSALInteractiveParams): Promise<MSALResult>;
-  acquireTokenSilent(params: MSALSilentParams): Promise<MSALResult>;
+  /**
+   * Acquire a token interactively
+   * @param {MSALInteractiveParams} params
+   * @return Result containing an access token and account identifier
+   * used for acquiring subsequent tokens silently
+   */
+  acquireToken(params: MSALInteractiveParams): Promise<MSALResult | undefined>;
+
+  /**
+   * Acquire a token silently
+   * @param {MSALSilentParams} params - Includes the account identifer retrieved from a
+   * previous interactive login
+   * @return Result containing an access token and account identifier
+   * used for acquiring subsequent tokens silently
+   */
+  acquireTokenSilent(params: MSALSilentParams): Promise<MSALResult | undefined>;
+
+  /**
+   * Get all accounts for which this application has refresh tokens
+   * @return Promise containing array of MSALAccount objects for which this application has refresh tokens.
+   */
   getAccounts(): Promise<MSALAccount[]>;
-  getAccount(accountIdentifier: string): Promise<MSALAccount>;
+
+  /**
+   * Retrieve the account matching the identifier
+   * @return Promise containing MSALAccount object
+   */
+  getAccount(accountIdentifier: string): Promise<MSALAccount | undefined>;
+
+  /**
+   * Removes all tokens from the cache for this application for the provided
+   * account.
+   * @param {MSALAccount} account
+   * @return A promise containing a boolean = true if account removal was successful
+   * otherwise rejects
+   */
   removeAccount(account: MSALAccount): Promise<boolean>;
+
+  /**
+   * Removes all tokens from the cache for this application for the provided
+   * account. Additionally, this will remove the account from the system browser.
+   * NOTE: iOS only. On Android and web this is the same as `removeAccount`.
+   * @param {MSALSignoutParams} params
+   * @return A promise which resolves if sign out is successful,
+   * otherwise rejects
+   * @platform ios
+   */
   signOut(params: MSALSignoutParams): Promise<boolean>;
 }
 
@@ -15,12 +56,41 @@ export interface MSALConfiguration {
     clientId: string;
     authority?: string;
     knownAuthorities?: string[];
+    /**
+     * If you are providing this property, you should probably use `Platform.select`,
+     * because the redirect uris will be different for each platform.
+     */
     redirectUri?: string;
   };
   /**
    * @platform web
    */
   cache?: Configuration['cache'] & { cacheLocation?: 'localStorage' | 'sessionStorage' };
+  /**
+   * Options as described here: {@link https://docs.microsoft.com/en-us/azure/active-directory/develop/msal-configuration}
+   * @platform android
+   */
+  androidConfigOptions?: MSALAndroidConfigOptions;
+}
+
+export interface MSALAndroidConfigOptions {
+  authorization_user_agent?: 'DEFAULT' | 'BROWSER' | 'WEBVIEW';
+  broker_redirect_uri_registered?: boolean;
+  browser_safelist?: {
+    browser_package_name: string;
+    browser_signature_hashes: string[];
+    browser_use_customTab: boolean;
+  }[];
+  http?: {
+    connect_timeout?: number;
+    read_timeout?: number;
+  };
+  logging?: {
+    pii_enabled?: boolean;
+    log_level?: 'ERROR' | 'WARNING' | 'INFO' | 'VERBOSE';
+    logcat_enabled?: boolean;
+  };
+  multiple_clouds_supported?: boolean;
 }
 
 export interface MSALInteractiveParams {